+++ .idea/workspace.xml
@@ -89,10 +89,11 @@
       <recent name="C:\2024-1_App1\client\front\assets\images" />
     </key>
     <key name="MoveFile.RECENT_KEYS">
-      <recent name="C:\Users\user\Desktop\wap\client\front\lib\main\errand_list\check_errand" />
-      <recent name="C:\Users\user\Desktop\wap\client\front\lib\main\errand_list" />
-      <recent name="C:\Users\user\Desktop\wap\client\front\lib\main" />
-      <recent name="C:\2024-1_App1\client\front\assets\images" />
+      <recent name="C:\Users\user\Desktop\wap\client\front\lib\main\errand_list\fix_errand" />
+      <recent name="C:\Users\user\Desktop\wap\client\front\lib" />
+      <recent name="C:\Users\user\Desktop\wap\client\front\lib\main\errand_list\check_errand\show_errand\show_errand_widget\table" />
+      <recent name="C:\Users\user\Desktop\wap\client\front\lib\main\errand_list\check_errand\show_errand\show_errand_widget\stamp\name_length" />
+      <recent name="C:\Users\user\Desktop\wap\client\front\lib\main\errand_list\check_errand\show_errand\show_errand_widget\stamp" />
     </key>
   </component>
   <component name="RunManager">
