+++ client/front/lib/main/errand_list/check_errand/show_errand/show_errand_widget/stamp/name_length/name_length5more.dart
