+++ client/front/lib/main/errand_list/check_errand/show_errand/show_errand_widget/table/table2.dart
