+++ client/front/lib/status_client.dart
@@ -12,7 +12,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 import 'main/errand_list/errand_list.dart';
 import 'map.dart';
-import 'show_errand/re-show_errand.dart';
+import 're-show_errand.dart';
 
 final storage = FlutterSecureStorage();
 class StatusContent{//진행중인 심부름이 간략하게 담고 있는 정보들
